+++ packages/plugins/AddressSearch/rollup.config.js
@@ -4,7 +4,6 @@ import scss from 'rollup-plugin-scss'
 import terser from '@rollup/plugin-terser'
 import typescript from 'rollup-plugin-typescript2'
 import vue from 'rollup-plugin-vue'
-import vuetify from 'rollup-plugin-vuetify'
 import { createFilter } from '@rollup/pluginutils'
 
 const external = createFilter(
@@ -40,7 +39,6 @@ export default {
     scss({ fileName: 'index.css', sourceMap: true }),
     // Types are not checked as typechecking does not work for vue-files with the current version, might be fixable with an upgrade of vue and vuex; see https://github.com/vuejs/vuex/pull/1121.
     typescript({ check: false }),
-    vuetify(),
     commonjs(),
     resolve(),
     terser(),