+++ packages/core/rollup.config.js
@@ -23,7 +23,6 @@ const external = createFilter(
     'olcs',
     'vue',
     'vuetify',
-    'vuetify/lib/components',
     'vuex',
   ],
   null,