+++ packages/core/src/components/MapContainer.vue
@@ -51,7 +51,6 @@ import Hammer from 'hammerjs'
 import i18next from 'i18next'
 import { defaults } from 'ol/interaction'
 import { LanguageOption, MoveHandleProperties } from '@polar/lib-custom-types'
-import { VApp } from 'vuetify/lib/components'
 import { SMALL_DISPLAY_HEIGHT, SMALL_DISPLAY_WIDTH } from '../utils/constants'
 import { addClusterStyle } from '../utils/addClusterStyle'
 import MapUi from './MapUi.vue'
@@ -63,7 +62,6 @@ export default Vue.extend({
   components: {
     MapUi,
     MoveHandle,
-    VApp,
   },
   props: {
     mapConfiguration: {