+++ packages/plugins/AddressSearch/src/components/AddressSearch.vue
@@ -14,7 +14,6 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import { VContainer, VToolbar } from 'vuetify/lib/components'
 import { mapGetters } from 'vuex'
 import GroupSelect from './GroupSelect.vue'
 import Results from './Results.vue'
@@ -26,8 +25,6 @@ export default Vue.extend({
     GroupSelect,
     SearchInput,
     Results,
-    VContainer,
-    VToolbar,
   },
   computed: {
     ...mapGetters(['clientWidth', 'hasSmallWidth', 'hasWindowSize']),